// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  agents: typeof agents;
  index: typeof index;
  prompts: typeof prompts;
  reviews: typeof reviews;
  rewrites: typeof rewrites;
  titles: typeof titles;
  transcripts: typeof transcripts;
  workflow: typeof workflow;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "internal">
>;